refactor clan member domain
